Convert popAndToast to typical script from module
To ease use, the module has been converted to a script revealing
popAndToast as a constant.
