+++ demo/popup.html
@@ -37,8 +37,8 @@
                 necessitatibus dolore distinctio iste quos minus ab quod fugit praesentium explicabo. Aut voluptatum
                 ullam et veniam maxime eligendi ducimus libero.</span></p>
     </main>
-    <script type="module" src="../src/pop-and-toast.mjs"></script>
-    <script type="module" src="main.js"></script>
+    <script src="../src/pop-and-toast.js"></script>
+    <script src="main.js"></script>
 </body>
 
 </html>
